Unified login/logout methods (#1111)
* Unified login/logout methods

* docstring

* documentation

* arrange imports

* add login reference

* fix

* disable codecov checks in diff page

* make interpreter_login official + doc
