+++ codecov.yml
@@ -3,3 +3,6 @@ comment:
   require_changes: true
   # https://docs.codecov.com/docs/pull-request-comments#after_n_builds
   after_n_builds: 11
+
+github_checks:
+  annotations: false
