+++ docs/source/_toctree.yml
@@ -1,28 +1,30 @@
 - sections:
-  - local: index
-    title: Home
-  - local: quick-start
-    title: Quick start
+    - local: index
+      title: Home
+    - local: quick-start
+      title: Quick start
   title: "Get started"
 - sections:
-  - local: how-to-manage
-    title: Create and manage repositories
-  - local: how-to-downstream
-    title: Download files from the Hub
-  - local: how-to-upstream
-    title: Upload files to the Hub
-  - local: searching-the-hub
-    title: Searching the Hub
-  - local: how-to-inference
-    title: Access the Inference API
-  - local: how-to-discussions-and-pull-requests
-    title: Interact with Discussions and Pull Requests
-  - local: how-to-cache
-    title: Manage the Cache
-  - local: how-to-model-cards
-    title: Create and Share Model Cards
+    - local: how-to-manage
+      title: Create and manage repositories
+    - local: how-to-downstream
+      title: Download files from the Hub
+    - local: how-to-upstream
+      title: Upload files to the Hub
+    - local: searching-the-hub
+      title: Searching the Hub
+    - local: how-to-inference
+      title: Access the Inference API
+    - local: how-to-discussions-and-pull-requests
+      title: Interact with Discussions and Pull Requests
+    - local: how-to-cache
+      title: Manage the Cache
+    - local: how-to-model-cards
+      title: Create and Share Model Cards
   title: "Guides"
 - sections:
+    - local: package_reference/login
+      title: Login and logout
     - local: package_reference/repository
       title: Managing local and online repositories
     - local: package_reference/hf_api
