+++ docs/source/how-to-upstream.mdx
@@ -15,14 +15,21 @@ Whenever you want to upload files to the Hub, you need to log in to your Hugging
    huggingface-cli login
    ```
 
-- Alternatively, if you prefer working from a Jupyter or Colaboratory notebook, log in with [`notebook_login`]:
+- Alternatively, you can programmatically login using [`login`] in a notebook or a script:
 
    ```python
-   >>> from huggingface_hub import notebook_login
-   >>> notebook_login()
+   >>> from huggingface_hub import login
+   >>> login()
    ```
 
-   [`notebook_login`] launches a widget in your notebook from which you can enter your Hugging Face credentials.
+   If ran in a Jupyter or Colaboratory notebook, [`login`] will launch a widget from
+   which you can enter your Hugging Face access token. Otherwise, a message will be
+   prompted in the terminal.
+
+   It is also possible to login programmatically without the widget by directly passing
+   the token to [`login`]. If you do so, be careful when sharing your notebook. It is
+   best practice to load the token from a secure vault instead of saving it in plain in
+   your Colaboratory notebook.
 
 ## Push files without Git
