+++ docs/source/package_reference/login.mdx
@@ -1,0 +1,20 @@
+# Login and logout
+
+The `huggingface_hub` library allows users to programmatically login and logout the
+machine to the Hub.
+
+## login
+
+[[autodoc]] login
+
+## interpreter_login
+
+[[autodoc]] interpreter_login
+
+## notebook_login
+
+[[autodoc]] notebook_login
+
+## logout
+
+[[autodoc]] logout
