+++ docs/source/quick-start.mdx
@@ -69,12 +69,11 @@ Once you have your User Access Token, run the following command in your terminal
 huggingface-cli login
 ```
 
-Or if you prefer to work from a Jupyter or Colaboratory notebook, then login with
-[`notebook_login`]:
+Or if you prefer to work from a Jupyter or Colaboratory notebook, then use [`login`]:
 
 ```py
->>> from huggingface_hub import notebook_login
->>> notebook_login()
+>>> from huggingface_hub import login
+>>> login()
 ```
 
 <Tip>
