+++ src/huggingface_hub/utils/__init__.py
@@ -54,8 +54,10 @@ from ._runtime import (
     get_torch_version,
     is_fastai_available,
     is_fastcore_available,
+    is_google_colab,
     is_graphviz_available,
     is_jinja_available,
+    is_notebook,
     is_pydot_available,
     is_tf_available,
     is_torch_available,
